# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sistema de llamado de mozos para restaurantes",
        "title": "Waiter API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/feedback": {
            "post": {
                "description": "Registra el feedback de un cliente sobre el servicio",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feedback"
                ],
                "summary": "Registrar feedback",
                "parameters": [
                    {
                        "description": "Datos del feedback",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecase.CreateFeedbackInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.Feedback"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/requests": {
            "post": {
                "description": "Crea una nueva solicitud de mesa (llamar mozo, pedir cuenta, ayuda)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "requests"
                ],
                "summary": "Crear solicitud",
                "parameters": [
                    {
                        "description": "Datos de la solicitud",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecase.CreateRequestInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.Request"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/requests/{requestId}": {
            "patch": {
                "description": "Actualiza el estado de una solicitud (marcar como atendida)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "requests"
                ],
                "summary": "Actualizar estado de solicitud",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la solicitud",
                        "name": "requestId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nuevo estado",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.updateRequestInput"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/restaurants": {
            "post": {
                "description": "Crea un nuevo restaurante",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "restaurants"
                ],
                "summary": "Crear restaurante",
                "parameters": [
                    {
                        "description": "Datos del restaurante",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecase.CreateRestaurantInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.Restaurant"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/restaurants/{restaurantId}": {
            "get": {
                "description": "Retorna un restaurante por su ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "restaurants"
                ],
                "summary": "Obtener restaurante",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del restaurante",
                        "name": "restaurantId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Restaurant"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/restaurants/{restaurantId}/requests/active": {
            "get": {
                "description": "Retorna las solicitudes activas de un restaurante",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "requests"
                ],
                "summary": "Obtener solicitudes activas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del restaurante",
                        "name": "restaurantId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Request"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/restaurants/{restaurantId}/tables": {
            "get": {
                "description": "Retorna las mesas de un restaurante",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tables"
                ],
                "summary": "Listar mesas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del restaurante",
                        "name": "restaurantId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Table"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Crea una nueva mesa para un restaurante",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tables"
                ],
                "summary": "Crear mesa",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del restaurante",
                        "name": "restaurantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos de la mesa",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.createTableInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.Table"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/tables/{tableId}/status": {
            "get": {
                "description": "Retorna las solicitudes de una mesa específica",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "requests"
                ],
                "summary": "Obtener estado de mesa",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la mesa",
                        "name": "tableId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Request"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Feedback": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "table_id": {
                    "type": "string"
                }
            }
        },
        "entity.Request": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/entity.RequestStatus"
                },
                "table_id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/entity.RequestType"
                }
            }
        },
        "entity.RequestStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "IN_PROCESS",
                "DONE"
            ],
            "x-enum-varnames": [
                "Pending",
                "InProcess",
                "Done"
            ]
        },
        "entity.RequestType": {
            "type": "string",
            "enum": [
                "CALL_WAITER",
                "ASK_BILL",
                "ASK_HELP"
            ],
            "x-enum-varnames": [
                "CallWaiter",
                "AskBill",
                "AskHelp"
            ]
        },
        "entity.Restaurant": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "plan": {
                    "type": "string"
                }
            }
        },
        "entity.Table": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "qr_code": {
                    "type": "string"
                },
                "restaurant_id": {
                    "type": "string"
                }
            }
        },
        "http.createTableInput": {
            "type": "object",
            "required": [
                "number"
            ],
            "properties": {
                "number": {
                    "type": "integer"
                }
            }
        },
        "http.updateRequestInput": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "$ref": "#/definitions/entity.RequestStatus"
                }
            }
        },
        "usecase.CreateFeedbackInput": {
            "type": "object",
            "required": [
                "score",
                "table_id"
            ],
            "properties": {
                "comment": {
                    "type": "string"
                },
                "score": {
                    "type": "integer",
                    "maximum": 5,
                    "minimum": 1
                },
                "table_id": {
                    "type": "string"
                }
            }
        },
        "usecase.CreateRequestInput": {
            "type": "object",
            "required": [
                "table_id",
                "type"
            ],
            "properties": {
                "table_id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/entity.RequestType"
                }
            }
        },
        "usecase.CreateRestaurantInput": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "plan": {
                    "type": "string"
                }
            }
        }
    }
}